requirements.txt
annotated-types==0.6.0
anyio==4.2.0
black==23.12.1
click==8.1.7
fastapi==0.109.0
fastapi-utils==0.2.1
flake8==7.0.0
h11==0.14.0
idna==3.6
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pathspec==0.12.1
platformdirs==4.1.0
pycodestyle==2.11.1
pydantic==1.10.14
pydantic_core==2.14.6
pyflakes==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.35.1
typing_extensions==4.9.0
tzdata==2023.4
uvicorn==0.27.0
